chore: move easter egg to different component
